backend/requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.2.0
python-dateutil==2.8.2
python-decouple==3.6
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
flask-cors